requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
idna==2.9
isort==4.3.21
pathspec==0.7.0
PyJWT==1.7.1
python-http-client==3.2.7
pytz==2019.3
regex==2020.4.4
requests==2.23.0
sendgrid==6.2.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
twilio==6.38.0
typed-ast==1.4.1
urllib3==1.25.8
